Compute a Probabilistic Boolean Network (PBN) for each cluster or each cell

After running performViper, this can be run to produce a PBN with the identity function probability for each node relative to the VIPER normalised enrichment score (NES) for cell. For computePbnByCluster, the median NES for all cells in that cluster will be used.

Usage

computePbnByCluster(ingres.object, range = c(-1, 1))

computePbnBySingleCell(ingres.object, range = c(-1, 1))

Arguments

ingres.object

An ingres object with a non-null VIPER slot.

range

A vector representing the range to which the NES are to be re-scaled. range[1]+range[2] == 0 should be TRUE. Defaults to [-1, 1]

Value

An ingres object with the cluster.pbn or the single.cell.pbn slot filled

Functions

computePbnByCluster: Compute a Probabilistic Boolean Network (PBN) for each cluster

computePbnBySingleCell: Compute a Probabilistic Boolean Network (PBN) for each cell

Examples

# Create an ingres object with viper slot
ing = createIngresObjectFromSeurat(
  small_blca_wang, "RNA", "data", network_genes, network
)
ing@viper = viper_results

ing = computePbnByCluster(ing)
head(ing@cluster.pbn)
#>   cluster  n          AKT        ATM    CHEK1_2     CyclinA    CyclinD1
#> 1       0 10 -0.097484176 -1.0000000  0.2166383  0.18131073  0.09277482
#> 2       1 10 -0.096326811 -1.0000000  0.2955960  0.35280697 -0.04390661
#> 3       2 10 -0.063815070 -0.8093695  0.2246400  1.00000000  0.00325166
#> 4       3 10 -0.063424756 -1.0000000  0.1330365  0.46560472  0.12554866
#> 5       4 10 -0.010348571 -0.9965950 -0.3066812 -0.18524741  0.30331777
#> 6       5 10  0.001876663 -0.5676380 -0.1472715 -0.01639584  0.28006367
#>      CyclinE1        E2F1        EGFR       FGFR3        GRB2         MDM2
#> 1  0.02914364 -0.18414952 -0.16332894 -0.16516364  0.26762876 -0.066543508
#> 2  0.04468990 -0.05342271 -0.16882291 -0.04755962  0.22798643 -0.084408374
#> 3  0.06342241  0.33638931 -0.13624438 -0.05261686  0.12801529  0.006134162
#> 4 -0.03427085  0.29593200 -0.41127203 -0.05913597 -0.01077991 -0.012822779
#> 5 -0.17783912 -0.69246793 -0.17103211 -0.07326116  0.15398728  0.089154534
#> 6 -0.01245400 -0.50915699 -0.07570176 -0.08304455  0.24640118 -0.028440399
#>         PI3K       PTEN       RAS        RB1        RBL2        SPRY
#> 1 -0.7496748 0.27021727 0.4981292 -0.1713536 -0.26610083 -0.11779371
#> 2 -0.7516899 0.26323344 0.5051441 -0.2130711 -0.34314007 -0.21140959
#> 3 -0.4454655 0.15474012 0.4328864 -0.2776757 -0.25838560 -0.17278461
#> 4 -0.6591261 0.02716736 0.4134223 -0.2151321 -0.26859445 -0.25136186
#> 5 -0.9429948 0.46758917 0.4085684  0.1744684  0.01932629  0.01295213
#> 6 -1.0000000 0.27611854 0.4314033  0.2048461 -0.19299611  0.06166337
#>           TP53    p14ARF    p21CIP
#> 1 -0.015769127 0.3315065 0.3546273
#> 2  0.005214565 0.3652086 0.2611126
#> 3 -0.059952999 0.4453124 0.1744224
#> 4 -0.011195217 0.3443604 0.4422260
#> 5 -0.026579701 0.1583249 1.0000000
#> 6 -0.016125642 0.2057767 0.7427967

ing = computePbnBySingleCell(ing)
head(ing@single.cell.pbn)
#>                         cell cluster         AKT        ATM      CHEK1_2
#> 1 sample1@AAGGAGCCAGCTTCGG-1       2  0.01817583 -0.7989694  0.099920471
#> 2 sample1@ACAGCTAAGATCCCGC-1       1 -0.07269741 -1.0000000  0.219362784
#> 3 sample1@ACGATGTAGTACATGA-1       3 -0.01603311 -0.9357372  0.513787703
#> 4 sample1@ACGGAGACAAATTGCC-1       2 -0.07899067 -0.7880775  0.055078000
#> 5 sample1@ACGGGCTCAAACCTAC-1       2 -0.04870924 -0.8023977 -0.002197616
#> 6 sample1@ACTGATGAGAGCTGGT-1       8  0.01102953 -0.6630933  0.153858311
#>     CyclinA    CyclinD1    CyclinE1        E2F1         EGFR       FGFR3
#> 1 1.0000000 0.005835620 -0.12590699  0.09456779 -0.002438333 -0.06437469
#> 2 0.2526539 0.100580602  0.03804537  0.14432723 -0.102653311 -0.08783756
#> 3 0.4896386 0.210391219 -0.04145495  1.00000000 -0.378800533  0.14514210
#> 4 1.0000000 0.060062911 -0.06046783 -0.06633384 -0.135412201 -0.03497538
#> 5 1.0000000 0.003268819  0.04396283  0.40529261  0.229916691 -0.24036193
#> 6 0.1336534 0.492846995 -0.26039627 -0.69295991 -0.509692556 -0.37852354
#>          GRB2         MDM2       PI3K        PTEN       RAS         RB1
#> 1  0.31113702 -0.033343239 -0.4071378  0.29342925 0.3769146 -0.02103803
#> 2  0.32854332 -0.094203638 -0.7358486  0.35724316 0.6074293 -0.15998474
#> 3 -0.07954781  0.054171975 -0.5368198 -0.05223321 0.5145371 -0.38879500
#> 4  0.14676515 -0.033723994 -0.5897442  0.20687463 0.5643470 -0.18693366
#> 5  0.05351330  0.008933491 -0.4302515  0.27558028 0.5007909 -0.38687170
#> 6  0.37222843 -0.326267067 -1.0000000  0.05096307 0.4309240  0.66462958
#>         RBL2       SPRY        TP53    p14ARF     p21CIP
#> 1 -0.2541966 -0.2188479 -0.10594942 0.3929228 0.38233341
#> 2 -0.2668762 -0.2855518  0.10555768 0.3691654 0.42156174
#> 3 -0.4278514 -0.3070425  0.15455175 0.5303340 0.01893101
#> 4 -0.2616696 -0.1796798 -0.09579184 0.4598230 0.29802852
#> 5 -0.3977755 -0.2124557  0.17860741 0.3479199 0.37278246
#> 6  0.3797944 -0.2995955 -0.12893251 0.1103229 0.37679082

# Restrict range to (-0.5, 0.5)
ing = computePbnByCluster(ing, range = c(-0.5, 0.5))
head(ing@cluster.pbn)
#>   cluster  n           AKT        ATM     CHEK1_2      CyclinA    CyclinD1
#> 1       0 10 -0.0487420878 -0.5000000  0.10831915  0.090655366  0.04638741
#> 2       1 10 -0.0481634055 -0.5000000  0.14779800  0.176403485 -0.02195330
#> 3       2 10 -0.0319075352 -0.4046847  0.11232000  0.500000000  0.00162583
#> 4       3 10 -0.0317123779 -0.5000000  0.06651825  0.232802361  0.06277433
#> 5       4 10 -0.0051742855 -0.4982975 -0.15334060 -0.092623707  0.15165888
#> 6       5 10  0.0009383313 -0.2838190 -0.07363577 -0.008197919  0.14003183
#>       CyclinE1        E2F1        EGFR       FGFR3         GRB2         MDM2
#> 1  0.014571819 -0.09207476 -0.08166447 -0.08258182  0.133814378 -0.033271754
#> 2  0.022344949 -0.02671136 -0.08441145 -0.02377981  0.113993213 -0.042204187
#> 3  0.031711206  0.16819465 -0.06812219 -0.02630843  0.064007646  0.003067081
#> 4 -0.017135423  0.14796600 -0.20563601 -0.02956798 -0.005389957 -0.006411390
#> 5 -0.088919561 -0.34623396 -0.08551606 -0.03663058  0.076993638  0.044577267
#> 6 -0.006227001 -0.25457850 -0.03785088 -0.04152228  0.123200589 -0.014220200
#>         PI3K       PTEN       RAS         RB1         RBL2         SPRY
#> 1 -0.3748374 0.13510864 0.2490646 -0.08567681 -0.133050414 -0.058896854
#> 2 -0.3758449 0.13161672 0.2525721 -0.10653556 -0.171570033 -0.105704793
#> 3 -0.2227327 0.07737006 0.2164432 -0.13883783 -0.129192799 -0.086392304
#> 4 -0.3295631 0.01358368 0.2067112 -0.10756603 -0.134297224 -0.125680931
#> 5 -0.4714974 0.23379459 0.2042842  0.08723422  0.009663146  0.006476063
#> 6 -0.5000000 0.13805927 0.2157017  0.10242307 -0.096498056  0.030831685
#>           TP53     p14ARF    p21CIP
#> 1 -0.007884563 0.16575326 0.1773137
#> 2  0.002607283 0.18260432 0.1305563
#> 3 -0.029976500 0.22265622 0.0872112
#> 4 -0.005597609 0.17218018 0.2211130
#> 5 -0.013289851 0.07916246 0.5000000
#> 6 -0.008062821 0.10288837 0.3713984

ing = computePbnBySingleCell(ing, range = c(-0.5, 0.5))
head(ing@single.cell.pbn)
#>                         cell cluster          AKT        ATM      CHEK1_2
#> 1 sample1@AAGGAGCCAGCTTCGG-1       2  0.009087914 -0.3994847  0.049960235
#> 2 sample1@ACAGCTAAGATCCCGC-1       1 -0.036348706 -0.5000000  0.109681392
#> 3 sample1@ACGATGTAGTACATGA-1       3 -0.008016553 -0.4678686  0.256893851
#> 4 sample1@ACGGAGACAAATTGCC-1       2 -0.039495337 -0.3940388  0.027539000
#> 5 sample1@ACGGGCTCAAACCTAC-1       2 -0.024354619 -0.4011988 -0.001098808
#> 6 sample1@ACTGATGAGAGCTGGT-1       8  0.005514765 -0.3315466  0.076929156
#>     CyclinA    CyclinD1    CyclinE1        E2F1         EGFR       FGFR3
#> 1 0.5000000 0.002917810 -0.06295350  0.04728390 -0.001219166 -0.03218735
#> 2 0.1263270 0.050290301  0.01902269  0.07216362 -0.051326656 -0.04391878
#> 3 0.2448193 0.105195610 -0.02072748  0.50000000 -0.189400267  0.07257105
#> 4 0.5000000 0.030031455 -0.03023391 -0.03316692 -0.067706100 -0.01748769
#> 5 0.5000000 0.001634409  0.02198142  0.20264630  0.114958345 -0.12018097
#> 6 0.0668267 0.246423498 -0.13019813 -0.34647995 -0.254846278 -0.18926177
#>          GRB2         MDM2       PI3K        PTEN       RAS         RB1
#> 1  0.15556851 -0.016671619 -0.2035689  0.14671463 0.1884573 -0.01051901
#> 2  0.16427166 -0.047101819 -0.3679243  0.17862158 0.3037147 -0.07999237
#> 3 -0.03977391  0.027085987 -0.2684099 -0.02611660 0.2572686 -0.19439750
#> 4  0.07338257 -0.016861997 -0.2948721  0.10343732 0.2821735 -0.09346683
#> 5  0.02675665  0.004466745 -0.2151258  0.13779014 0.2503954 -0.19343585
#> 6  0.18611422 -0.163133533 -0.5000000  0.02548154 0.2154620  0.33231479
#>         RBL2        SPRY        TP53     p14ARF      p21CIP
#> 1 -0.1270983 -0.10942395 -0.05297471 0.19646142 0.191166703
#> 2 -0.1334381 -0.14277588  0.05277884 0.18458272 0.210780872
#> 3 -0.2139257 -0.15352125  0.07727588 0.26516699 0.009465505
#> 4 -0.1308348 -0.08983989 -0.04789592 0.22991148 0.149014262
#> 5 -0.1988878 -0.10622785  0.08930371 0.17395994 0.186391231
#> 6  0.1898972 -0.14979777 -0.06446626 0.05516143 0.188395408